asdsad

+++ C07/ex04/ft_convert_base2.c
@@ -6,7 +6,7 @@
 /*   By: asierra <asierra@student.42malaga.com>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/08/06 14:13:04 by asierra           #+#    #+#             */
-/*   Updated: 2026/08/07 12:21:39 by asierra          ###   ########.fr       */
+/*   Updated: 2026/08/07 12:53:12 by asierra          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 int		ft_is_space(int c);
 int		ft_strlen(char *str);
 int		is_valid_base(char *base);
-int		convert(char *str, char *base);
+long	convert(char *str, char *base);
 long	ft_atoi_base(char *str, char *base);
 
 int	ft_base_number_len(long nbr, int base_len)
@@ -82,8 +82,9 @@ char	*ft_convert_base(char *nbr, char *base_from, char *base_to)
 	ft_fill_char(nbr_int, base_to, base_to_size, conversion);
 	return (conversion);
 }
-/* 
+/*
 #include <stdio.h>
+
 int main (int argc, char **argv)
 {
 	char	*conversion;
